// src/src/lib/database.types.ts
export interface Database {
  public: {
    Tables: {
// ...
      categories: {
        Row: {
          id: string
          name: string
          slug: string
          created_at: string
          icon: string | null
          description: string | null
          image: string | null
        }
        Insert: {
          id?: string
          name: string
          slug: string
          created_at?: string
          icon?: string | null
          description?: string | null
          image?: string | null
        }
        Update: {
          name?: string
          slug?: string
          icon?: string | null
          description?: string | null
          image?: string | null
        }
        Relationships: []
      }
// ...
    }

    Views: Record<string, never>
    Functions: Record<string, never>
    Enums: Record<string, never>
    CompositeTypes: Record<string, never>
  }
}

export type ProfileRow = Database['public']['Tables']['profiles']['Row']
export type CategoryRow = Database['public']['Tables']['categories']['Row']